Simplify InitStatus and use class syntax
This was necessary to get the docstring to be picked up by mkdocstrings.

The disadvantage is that all Xilinx errors get mapped to the same enum
value, but this is probably not a significant drawback, given that
manufacturer help would most likely be needed to troubleshoot any of
these errors anyway.